Move `publishers.exception.EventNotFound` to `storage.query.exceptions`
