+++ mobilizon_reshare/storage/query/exceptions.py
@@ -1,0 +1,6 @@
+class EventNotFound(Exception):
+    """Event is not present in the database"""
+
+
+class DuplicateEvent(ValueError):
+    """A duplicate mobilizon_id has been found in the database"""
